bazel: untangle the wasm build
Untangle the wasm build so that tests can depend on private headers.
We're not fully running the tests yet because we'll need to hook the
tinygo compiler up into the build system first.
