+++ src/v/redpanda/BUILD
@@ -59,7 +59,9 @@ redpanda_cc_library(
         "//src/v/utils:stop_signal",
         "//src/v/utils:uuid",
         "//src/v/version",
-        "//src/v/wasm",
+        "//src/v/wasm:api",
+        "//src/v/wasm:cache",
+        "//src/v/wasm:impl",
         "@abseil-cpp//absl/log:globals",
         "@boost//:program_options",
         "@protobuf",
