+++ src/v/transform/BUILD
@@ -87,7 +87,10 @@ redpanda_cc_library(
         "//src/v/transform/rpc",
         "//src/v/utils:human",
         "//src/v/utils:prefix_logger",
-        "//src/v/wasm",
+        "//src/v/wasm:api",
+        "//src/v/wasm:cache",
+        "//src/v/wasm:transform_probe",
+        "//src/v/wasm:wasi_logger",
         "@abseil-cpp//absl/algorithm:container",
         "@abseil-cpp//absl/container:btree",
         "@abseil-cpp//absl/container:flat_hash_map",
