+++ src/v/wasm/errc.h
