+++ src/v/wasm/fwd.h
