+++ src/v/wasm/impl.cc
@@ -1,0 +1,23 @@
+/*
+ * Copyright 2023 Redpanda Data, Inc.
+ *
+ * Use of this software is governed by the Business Source License
+ * included in the file licenses/BSL.md
+ *
+ * As of the Change Date specified in that file, in accordance with
+ * the Business Source License, use of this software will be governed
+ * by the Apache License, Version 2.0
+ */
+
+#include "wasm/impl.h"
+
+#include "wasm/schema_registry.h"
+#include "wasm/wasmtime.h"
+
+namespace wasm {
+std::unique_ptr<runtime>
+create_default_runtime(pandaproxy::schema_registry::api* schema_reg) {
+    return wasmtime::create_runtime(
+      wasm::schema_registry::make_default(schema_reg));
+}
+} // namespace wasm
