+++ src/v/wasm/transform_probe.cc
@@ -11,7 +11,6 @@
 
 #include "wasm/transform_probe.h"
 
-#include "config/configuration.h"
 #include "metrics/prometheus_sanitize.h"
 #include "utils/log_hist.h"
